+++ apps/modernjs-ssr/dynamic-remote/modern.config.ts
@@ -5,6 +5,8 @@ import { moduleFederationPlugin } from '@module-federation/modern-js';
 export default defineConfig({
   dev: {
     port: 3053,
+    // FIXME: it should be removed , related issue: https://github.com/web-infra-dev/modern.js/issues/5999
+    host: '0.0.0.0',
   },
   runtime: {
     router: true,