+++ packages/runtime/src/utils/preload.ts
@@ -212,6 +212,7 @@ export function preloadAssets(
           cb: () => {},
           attrs: {
             fetchpriority: 'high',
+            type: remoteInfo?.type === 'module' ? 'module' : 'text/javascript',
           },
           createScriptHook: (url: string, attrs: any) => {
             const res = host.loaderHook.lifecycle.createScript.emit({