+++ packages/bridge/bridge-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @module-federation/bridge-react
 
+## 0.3.3
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.3.3
+
+## 0.3.2
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.3.2
+
 ## 0.3.1
 
 ### Patch Changes