+++ packages/enhanced/CHANGELOG.md
@@ -1,5 +1,32 @@
 # [0.2.0-canary.5](https://github.com/module-federation/core/compare/enhanced-0.2.0-canary.4...enhanced-0.2.0-canary.5) (2023-11-20)
 
+## 0.3.3
+
+### Patch Changes
+
+- @module-federation/rspack@0.3.3
+- @module-federation/sdk@0.3.3
+- @module-federation/runtime-tools@0.3.3
+- @module-federation/managers@0.3.3
+- @module-federation/manifest@0.3.3
+- @module-federation/dts-plugin@0.3.3
+- @module-federation/bridge-react-webpack-plugin@0.3.3
+
+## 0.3.2
+
+### Patch Changes
+
+- 85ae159: feat: support rspack ssr
+- Updated dependencies [0de1c83]
+- Updated dependencies [85ae159]
+  - @module-federation/dts-plugin@0.3.2
+  - @module-federation/manifest@0.3.2
+  - @module-federation/rspack@0.3.2
+  - @module-federation/sdk@0.3.2
+  - @module-federation/runtime-tools@0.3.2
+  - @module-federation/managers@0.3.2
+  - @module-federation/bridge-react-webpack-plugin@0.3.2
+
 ## 0.3.1
 
 ### Patch Changes